ProductDataFixtureReferenceInjector moved to framework DataFixtures folder
- ProductDataFixtureReferenceInjector is specific for framework and it does not match with Components definition
